+++ design-system/icon/src/components/__tests__/unit/server-side-hydrate.tsx
@@ -12,7 +12,7 @@ afterEach(() => {
 });
 
 test('should ssr then hydrate icon correctly', async () => {
-  const [example] = await getExamplesFor('icon');
+  const [example] = await getExamplesFor('@atlaskit/icon');
   const Example = require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
   const elem = document.createElement('div');
   elem.innerHTML = await ssr(example.filePath);