+++ editor/embedded-document/src/__tests__/unit/service-provider.test.ts
@@ -1,4 +1,4 @@
-import fetchMock from 'fetch-mock/src/client';
+import fetchMock from 'fetch-mock/cjs/client';
 import ServiceProvider from '../../provider/service-provider';
 import {
   docId,