+++ media/media-picker/src/popup/tools/fetcher/__tests__/unit/fetcher.spec.tsx
@@ -1,4 +1,4 @@
-import fetchMock from 'fetch-mock';
+import fetchMock from 'fetch-mock/cjs/client';
 import { Service } from '../../../../domain';
 import {
   MediaApiFetcher,